decouple more things
EMPTY_BYTE_ARRAY;
getSeed;
move src test Utils things into src test TestUtils;
