+++ src/test/java/com/esaulpaugh/headlong/abi/TupleTest.java
@@ -31,7 +31,7 @@ public class TupleTest {
     @Test
     public void testTypeSafety() throws Throwable {
 
-        Random rand = new Random(MonteCarloTest.getSeed(System.nanoTime()));
+        Random rand = new Random(TestUtils.getSeed(System.nanoTime()));
 
         for (int i = 0; i < 1000; i++) {
 
@@ -66,7 +66,7 @@ public class TupleTest {
 
     @Test
     public void fuzzNulls() throws Throwable {
-        Random r = new Random(MonteCarloTest.getSeed(System.nanoTime()));
+        Random r = new Random(TestUtils.getSeed(System.nanoTime()));
         for (int i = 0; i < 1000; i++) {
             MonteCarloTestCase mctc = new MonteCarloTestCase(r.nextLong());
             Tuple args = mctc.argsTuple;
@@ -119,7 +119,7 @@ public class TupleTest {
 
         final int len = master.length;
 
-        Random rand = new Random(MonteCarloTest.getSeed(System.nanoTime()));
+        Random rand = new Random(TestUtils.getSeed(System.nanoTime()));
 
         shuffle(master, rand);
