+++ src/test/java/com/esaulpaugh/headlong/util/StringsTest.java
@@ -15,7 +15,7 @@
 */
 package com.esaulpaugh.headlong.util;
 
-import com.esaulpaugh.headlong.abi.MonteCarloTest;
+import com.esaulpaugh.headlong.TestUtils;
 import com.migcomponents.migbase64.Base64;
 import org.junit.jupiter.api.Test;
 
@@ -28,7 +28,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class StringsTest {
 
-    private static final Random RAND = new Random(MonteCarloTest.getSeed(System.nanoTime()));
+    private static final Random RAND = new Random(TestUtils.getSeed(System.nanoTime()));
 
     private static final Supplier<byte[]> SUPPLY_RANDOM = () -> {
         byte[] x = new byte[RAND.nextInt(115)];
@@ -64,7 +64,7 @@ public class StringsTest {
 
     @Test
     public void base64NoOptions() {
-        Random rand = new Random(MonteCarloTest.getSeed(System.nanoTime()));
+        Random rand = new Random(TestUtils.getSeed(System.nanoTime()));
         java.util.Base64.Encoder mimeEncoder = java.util.Base64.getMimeEncoder();
         java.util.Base64.Decoder mimeDecoder = java.util.Base64.getMimeDecoder();
         for(int j = 0; j < 250; j++) {
@@ -80,7 +80,7 @@ public class StringsTest {
 
     @Test
     public void base64PaddedNoLineSep() {
-        Random rand = new Random(MonteCarloTest.getSeed(System.nanoTime()));
+        Random rand = new Random(TestUtils.getSeed(System.nanoTime()));
         java.util.Base64.Encoder encoder = java.util.Base64.getUrlEncoder();
         for (int j = 0; j < 250; j++) {
             byte[] x = new byte[j];
@@ -95,7 +95,7 @@ public class StringsTest {
 
     @Test
     public void base64Default() {
-        Random rand = new Random(MonteCarloTest.getSeed(System.nanoTime()));
+        Random rand = new Random(TestUtils.getSeed(System.nanoTime()));
         for(int j = 3; j < 250; j++) {
             byte[] x = new byte[j];
             rand.nextBytes(x);
